Add files via upload
This program aims to solve recurrent problems in child abduction at maternity facilities. In this I make use of LBPH to match the baby to the mother. This can be used once the mother checks out of the hospital. If a stranger comes into the facility and tries to steal the baby, an alarm is sounded so that the security takes note. Also, the stranger's images are captured by the webcam and stored in  a database for reference.